+++ typescript/packages/subsurface-viewer/src/layers/wells/layers/wellLabelLayer.ts
@@ -1,31 +1,14 @@
-import type {
-    Color,
-    DefaultProps,
-    Position,
-    UpdateParameters,
-} from "@deck.gl/core";
+import type { DefaultProps, Position, UpdateParameters } from "@deck.gl/core";
 import { OrbitViewport } from "@deck.gl/core";
 import type { TextLayerProps } from "@deck.gl/layers";
 import { TextLayer } from "@deck.gl/layers";
-import type {
-    Feature,
-    GeoJsonProperties,
-    Geometry,
-    GeometryCollection,
-    LineString,
-} from "geojson";
+import type { Feature, GeoJsonProperties, Geometry } from "geojson";
 import _ from "lodash";
 import { Vector2, Vector3 } from "math.gl";
 import type { Position3D } from "../../utils/layerTools";
+import { getTrajectory } from "../utils/trajectory";
 
-type NumberPair = [number, number];
-type StyleData = NumberPair | Color | number;
-type StyleAccessorFunction = (
-    object: Feature,
-    objectInfo?: Record<string, unknown>
-) => StyleData;
 type WellLabelLayerData = Feature<Geometry, GeoJsonProperties>;
-type ColorAccessor = Color | StyleAccessorFunction | undefined;
 
 export enum LabelOrientation {
     HORIZONTAL = "horizontal",
@@ -54,56 +37,6 @@ const DEFAULT_PROPS: DefaultProps<WellLabelLayerProps> = {
     orientation: LabelOrientation.HORIZONTAL,
     backgroundPadding: [5, 1, 5, 1],
 };
-
-function getLineStringGeometry(well_object: Feature): LineString {
-    return (well_object.geometry as GeometryCollection)?.geometries.find(
-        (item: { type: string }) => item.type == "LineString"
-    ) as LineString;
-}
-
-function getColor(accessor: ColorAccessor) {
-    if (accessor as Color) {
-        return accessor as Color;
-    }
-
-    return (object: Feature, objectInfo?: Record<string, unknown>): Color => {
-        if (typeof accessor === "function") {
-            const color = (accessor as StyleAccessorFunction)(
-                object,
-                objectInfo
-            ) as Color;
-            if (color) {
-                return color;
-            }
-        }
-        return object.properties?.["color"] as Color;
-    };
-}
-
-// return trajectory visibility based on alpha of trajectory color
-function isTrajectoryVisible(
-    well_object: Feature,
-    color_accessor: ColorAccessor
-): boolean {
-    let alpha;
-    const accessor = getColor(color_accessor);
-    if (typeof accessor === "function") {
-        alpha = accessor(well_object)?.[3];
-    } else {
-        alpha = (accessor as Color)?.[3];
-    }
-    return alpha !== 0;
-}
-
-// Return Trajectory data from LineString Geometry if it's visible (checking trajectory visiblity based on line color)
-function getTrajectory(
-    well_object: Feature,
-    color_accessor: ColorAccessor
-): Position[] | undefined {
-    if (isTrajectoryVisible(well_object, color_accessor))
-        return getLineStringGeometry(well_object)?.coordinates as Position[];
-    else return undefined;
-}
 
 export class WellLabelLayer extends TextLayer<
     WellLabelLayerData,
