+++ typescript/packages/subsurface-viewer/src/layers/wells/wellsLayer.ts
@@ -26,7 +26,6 @@ import type {
     GeoJsonProperties,
     Geometry,
     GeometryCollection,
-    LineString,
     Point,
 } from "geojson";
 import { distance, dot, subtract } from "mathjs";
@@ -53,17 +52,12 @@ import {
     invertPath,
     splineRefine,
 } from "./utils/spline";
+import type { NumberPair, StyleAccessorFunction } from "../types";
+import { getColor, getTrajectory } from "./utils/trajectory";
 
-type StyleAccessorFunction = (
-    object: Feature,
-    objectInfo?: Record<string, unknown>
-) => StyleData;
-
-type NumberPair = [number, number];
 type DashAccessor = boolean | NumberPair | StyleAccessorFunction | undefined;
 type ColorAccessor = Color | StyleAccessorFunction | undefined;
 type SizeAccessor = number | StyleAccessorFunction | undefined;
-type StyleData = NumberPair | Color | number;
 
 type LineStyleAccessor = {
     color?: ColorAccessor;
@@ -253,25 +247,6 @@ function getDashFactor(
         } else {
             return multiply(DEFAULT_DASH, factor);
         }
-    };
-}
-
-function getColor(accessor: ColorAccessor) {
-    if (accessor as Color) {
-        return accessor as Color;
-    }
-
-    return (object: Feature, objectInfo?: Record<string, unknown>): Color => {
-        if (typeof accessor === "function") {
-            const color = (accessor as StyleAccessorFunction)(
-                object,
-                objectInfo
-            ) as Color;
-            if (color) {
-                return color;
-            }
-        }
-        return object.properties?.["color"] as Color;
     };
 }
 
@@ -869,40 +844,9 @@ function getPointGeometry(well_object: Feature): Point {
     ) as Point;
 }
 
-function getLineStringGeometry(well_object: Feature): LineString {
-    return (well_object.geometry as GeometryCollection)?.geometries.find(
-        (item: { type: string }) => item.type == "LineString"
-    ) as LineString;
-}
-
 // Return well head position from Point Geometry
 function getWellHeadPosition(well_object: Feature): Position {
     return getPointGeometry(well_object)?.coordinates as Position;
-}
-
-// return trajectory visibility based on alpha of trajectory color
-function isTrajectoryVisible(
-    well_object: Feature,
-    color_accessor: ColorAccessor
-): boolean {
-    let alpha;
-    const accessor = getColor(color_accessor);
-    if (typeof accessor === "function") {
-        alpha = accessor(well_object)?.[3];
-    } else {
-        alpha = (accessor as Color)?.[3];
-    }
-    return alpha !== 0;
-}
-
-// Return Trajectory data from LineString Geometry if it's visible (checking trajectory visiblity based on line color)
-function getTrajectory(
-    well_object: Feature,
-    color_accessor: ColorAccessor
-): Position[] | undefined {
-    if (isTrajectoryVisible(well_object, color_accessor))
-        return getLineStringGeometry(well_object)?.coordinates as Position[];
-    else return undefined;
 }
 
 function getWellMds(well_object: Feature): number[] {
